Move task to folder and separate runner into separate module
